# docs/swagger.yaml
basePath: /
definitions:
  github_com_torchlabssoftware_subnetwork_system_internal_server_models.CreateUserRequest:
    properties:
      allow_pools:
        items:
          $ref: '#/definitions/github_com_torchlabssoftware_subnetwork_system_internal_server_models.PoolDataStat'
        type: array
      ip_whitelist:
        items:
          type: string
        type: array
    type: object
  github_com_torchlabssoftware_subnetwork_system_internal_server_models.CreateUserResponce:
    properties:
      allow_pools:
        items:
          type: string
        type: array
      created_at:
        type: string
      id:
        type: string
      ip_whitelist:
        items:
          type: string
        type: array
      password:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_torchlabssoftware_subnetwork_system_internal_server_models.PoolDataStat:
    properties:
      data_limit:
        type: integer
      data_usage:
        type: integer
      pool:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Server for Subnetwork System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Subnetwork System API
  version: "1.0"
paths:
  /admin/users:
    post:
      consumes:
      - application/json
      description: Creates a user with optional IP whitelist and pool allowances
      parameters:
      - description: Create User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_torchlabssoftware_subnetwork_system_internal_server_models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_torchlabssoftware_subnetwork_system_internal_server_models.CreateUserResponce'
        "400":
          description: Invalid request body
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Create a new user
      tags:
      - users
swagger: "2.0"